Remove redundant Fraction class
